requirements.txt
fastapi
fastapi-restful
uvicorn
fsspec
requests
python-podcastindex